+++ mappers/Mapper002.h
@@ -12,14 +12,14 @@ public:
 	std::string name() const override;
 
 public:
-	void write_8(uint16_t address, uint8_t value) override;
-	void write_9(uint16_t address, uint8_t value) override;
-	void write_a(uint16_t address, uint8_t value) override;
-	void write_b(uint16_t address, uint8_t value) override;
-	void write_c(uint16_t address, uint8_t value) override;
-	void write_d(uint16_t address, uint8_t value) override;
-	void write_e(uint16_t address, uint8_t value) override;
-	void write_f(uint16_t address, uint8_t value) override;
+	void write_8(uint32_t address, uint8_t value) override;
+	void write_9(uint32_t address, uint8_t value) override;
+	void write_a(uint32_t address, uint8_t value) override;
+	void write_b(uint32_t address, uint8_t value) override;
+	void write_c(uint32_t address, uint8_t value) override;
+	void write_d(uint32_t address, uint8_t value) override;
+	void write_e(uint32_t address, uint8_t value) override;
+	void write_f(uint32_t address, uint8_t value) override;
 
 private:
 	uint8_t chr_ram_[0x2000] = {};
