+++ mappers/Mapper003.cpp
@@ -25,7 +25,7 @@ std::string Mapper3::name() const {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper3::write_8(uint16_t address, uint8_t value) {
+void Mapper3::write_8(uint32_t address, uint8_t value) {
 	(void)address;
 	set_chr_0000_1fff(value);
 }
@@ -33,7 +33,7 @@ void Mapper3::write_8(uint16_t address, uint8_t value) {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper3::write_9(uint16_t address, uint8_t value) {
+void Mapper3::write_9(uint32_t address, uint8_t value) {
 	(void)address;
 	set_chr_0000_1fff(value);
 }
@@ -41,7 +41,7 @@ void Mapper3::write_9(uint16_t address, uint8_t value) {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper3::write_a(uint16_t address, uint8_t value) {
+void Mapper3::write_a(uint32_t address, uint8_t value) {
 	(void)address;
 	set_chr_0000_1fff(value);
 }
@@ -49,7 +49,7 @@ void Mapper3::write_a(uint16_t address, uint8_t value) {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper3::write_b(uint16_t address, uint8_t value) {
+void Mapper3::write_b(uint32_t address, uint8_t value) {
 	(void)address;
 	set_chr_0000_1fff(value);
 }
@@ -57,7 +57,7 @@ void Mapper3::write_b(uint16_t address, uint8_t value) {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper3::write_c(uint16_t address, uint8_t value) {
+void Mapper3::write_c(uint32_t address, uint8_t value) {
 	(void)address;
 	set_chr_0000_1fff(value);
 }
@@ -65,7 +65,7 @@ void Mapper3::write_c(uint16_t address, uint8_t value) {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper3::write_d(uint16_t address, uint8_t value) {
+void Mapper3::write_d(uint32_t address, uint8_t value) {
 	(void)address;
 	set_chr_0000_1fff(value);
 }
@@ -73,7 +73,7 @@ void Mapper3::write_d(uint16_t address, uint8_t value) {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper3::write_e(uint16_t address, uint8_t value) {
+void Mapper3::write_e(uint32_t address, uint8_t value) {
 	(void)address;
 	set_chr_0000_1fff(value);
 }
@@ -81,7 +81,7 @@ void Mapper3::write_e(uint16_t address, uint8_t value) {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper3::write_f(uint16_t address, uint8_t value) {
+void Mapper3::write_f(uint32_t address, uint8_t value) {
 	(void)address;
 	set_chr_0000_1fff(value);
 }
