+++ mappers/Mapper033.cpp
@@ -22,35 +22,35 @@ std::string Mapper33::name() const {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper33::write_8(uint16_t address, uint8_t value) {
+void Mapper33::write_8(uint32_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper33::write_9(uint16_t address, uint8_t value) {
+void Mapper33::write_9(uint32_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper33::write_a(uint16_t address, uint8_t value) {
+void Mapper33::write_a(uint32_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper33::write_b(uint16_t address, uint8_t value) {
+void Mapper33::write_b(uint32_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper33::write_handler(uint16_t address, uint8_t value) {
+void Mapper33::write_handler(uint32_t address, uint8_t value) {
 
 	switch (address & 0xa003) {
 	case 0x8000:
