+++ mappers/Mapper033.h
@@ -11,13 +11,13 @@ public:
 public:
 	std::string name() const override;
 
-	void write_8(uint16_t address, uint8_t value) override;
-	void write_9(uint16_t address, uint8_t value) override;
-	void write_a(uint16_t address, uint8_t value) override;
-	void write_b(uint16_t address, uint8_t value) override;
+	void write_8(uint32_t address, uint8_t value) override;
+	void write_9(uint32_t address, uint8_t value) override;
+	void write_a(uint32_t address, uint8_t value) override;
+	void write_b(uint32_t address, uint8_t value) override;
 
 private:
-	void write_handler(uint16_t address, uint8_t value);
+	void write_handler(uint32_t address, uint8_t value);
 };
 
 #endif
