+++ mappers/Mapper040.h
@@ -11,18 +11,18 @@ public:
 public:
 	std::string name() const override;
 
-	void write_8(uint16_t address, uint8_t value) override;
-	void write_9(uint16_t address, uint8_t value) override;
-	void write_a(uint16_t address, uint8_t value) override;
-	void write_b(uint16_t address, uint8_t value) override;
-	void write_e(uint16_t address, uint8_t value) override;
-	void write_f(uint16_t address, uint8_t value) override;
+	void write_8(uint32_t address, uint8_t value) override;
+	void write_9(uint32_t address, uint8_t value) override;
+	void write_a(uint32_t address, uint8_t value) override;
+	void write_b(uint32_t address, uint8_t value) override;
+	void write_e(uint32_t address, uint8_t value) override;
+	void write_f(uint32_t address, uint8_t value) override;
 
 public:
 	void cpu_sync() override;
 
 private:
-	uint16_t irq_counter_ = 0;
+	uint32_t irq_counter_ = 0;
 	bool irq_enabled_     = false;
 };
